Fixed bugs in enableCheck().
1. If the current state is the new state given as the argument to
enableCheck(), don't do anything.

2. if we're enabling checks, a recomputation of the crc values
shouldn't be caused at the time enableCheck() is called because the
value it gets may be wrong because of the time when it's usually
called.  It's usually called in the same function that issued a reload
to the file, and it's apparently possible for the data to be either
incomplete or wrong or *something*.  It's better not to reload during
that time, and instead let a cycle pass.