More efficient scheme allows us to increase the check frequency.
The code in crc-check.js and the back-end handler now provide an MD5
hash, so that the checker can transfer that each time instead of the
whole file every time.  This means a lot less data is being passed back
and forth, which means we can afford to increase the check frequency.