Use an md5 file on the server, CRC only as a fallback.
This now tries to download an MD5 has file from the server 1st, so that
the checker can transfer that each time instead of the whole file every
time.  This means a lot less data is being passed back and forth, which
means we can afford to increase the check frequency.